requirements.txt
appdirs==1.4.3
asgiref==3.2.3
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
Django==3.0
django-import-export==2.0.2
et-xmlfile==1.0.1
filelock==3.0.12
importlib-metadata==1.6.0
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
pytz==2019.3
PyYAML==5.3.1
six==1.12.0
sqlparse==0.3.0
tablib==1.1.0
virtualenv==20.0.18
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
